Reading IR sensors and publishing to CAN Bus.
This script runs on the CAN Node of the race car. It gathers the data emitted via I2C from the infrared temperature sensors (MLX9621) located on the brake discs and tyres. Afterwards it composes the respective CAN messages and publishes the values on the CAN Bus to be read from the Telemetry system.